Refatoracao do codigo
