+++ app/src/main/AndroidManifest.xml
@@ -15,12 +15,12 @@
         android:theme="@style/Theme.BitCoin_Price"
         tools:targetApi="31">
         <activity
-            android:name=".Homescreen"
+            android:name=".ui.main.MainActivity"
             android:exported="false"
             android:screenOrientation="portrait"/>
         <activity
 
-            android:name=".MainActivity"
+            android:name=".ui.splash.SplashScreen"
             android:exported="true"
             android:screenOrientation="portrait">
             <intent-filter>
