--- app/src/main/java/com/example/bitcoin_price/MarketPriceMappers.kt
@@ -1,12 +1,0 @@
-package com.example.bitcoin_price
-
-import com.example.bitcoin_price.data.MarketPriceValue
-import com.example.bitcoin_price.data.MarketPriceValueEntity
-
-fun MarketPriceValue.toEntity(): MarketPriceValueEntity {
-    return MarketPriceValueEntity(
-        timestamp = this.x,
-        price = this.y
-    )
-}
-
