+++ app/src/main/java/com/example/bitcoin_price/data/local/MarketPriceDao.kt
@@ -1,4 +1,4 @@
-package com.example.bitcoin_price.data
+package com.example.bitcoin_price.data.local
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
@@ -6,12 +6,13 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 
+
 @Dao
 interface MarketPriceDao {
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertAll(prices : List<MarketPriceValueEntity>)
+    @Insert(onConflict = OnConflictStrategy.Companion.REPLACE)
+    suspend fun insertAll(prices: List<MarketPriceValueEntity>)
 
     @Query("SELECT * FROM market_price_values")
-     fun getAll(): LiveData<List<MarketPriceValueEntity>>
+    fun getAll(): LiveData<List<MarketPriceValueEntity>>
 
 }
