+++ app/src/main/java/com/example/bitcoin_price/data/remote/RetrofitClient.kt
@@ -1,9 +1,7 @@
-package com.example.bitcoin_price.data
+package com.example.bitcoin_price.data.remote
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-
-// Responsável pela comunicaçao com a API do Blockchain. Converte o JSON para um objeto Kotlin.
 
 object RetrofitClient {
     private const val BASE_URL = "https://api.blockchain.info"
