+++ app/src/main/res/values-night/colors.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="background_button">#000000</color>
+    <color name="button_text_color">#000000</color>
     <color name="text_color">#FFFFFF</color>
     <color name="border_color">#444444</color>
     <color name="selected_button">#FFFFFF</color>
