+++ app/src/main/res/values/colors.xml
@@ -4,6 +4,7 @@
     <color name="black">#000000</color>
     <color name="transparent">#00A8A8A8</color>
         <color name="background_button">#FFFFFF</color>
+        <color name="button_text_color">#FFFFFF</color>
         <color name="text_color">#000000</color>
         <color name="border_color">#CCCCCC</color>
         <color name="selected_button">#000000</color>
